requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
httpx==0.25.0
transformers==4.35.2
torch==2.1.0
pillow==10.1.0
langchain==0.0.335
openai==1.3.5
google-cloud-storage==2.13.0
uuid==1.30
tenacity==8.2.3
pytest==7.4.3
black==23.11.0
isort==5.12.0
flake8==6.1.0
loguru==0.7.2 